Extract methylation or coverage matrices

Extract methylation or coverage matrices

get_matrix(m, type = "M", add_loci = FALSE, in_granges = FALSE)

Arguments

m
methrix object

type
can be M or C. Default 'M'

add_loci
Default FALSE. If TRUE adds CpG position info to the matrix and returns as a data.table

in_granges
Do you want the outcome in GRanges?

Value

Coverage or Methylation matrix

Details

Takes methrix object and returns user specified methylation or coverage matrix

Examples

data('methrix_data')
#Get methylation matrix
get_matrix(m = methrix_data, type = 'M')

#>               C1         C2        N1         N2
#>   [1,] 0.1538462 0.28571429 0.5555556 0.30000000
#>   [2,]        NA 0.50000000 0.0000000         NA
#>   [3,] 0.5555556 0.70000000 0.3333333 0.80000000
#>   [4,] 0.1818182 0.25000000 0.5833333 0.25000000
#>   [5,] 0.6666667 1.00000000 0.8823529 0.87500000
#>   [6,] 0.8461538 1.00000000 0.8333333 0.92857143
#>   [7,] 0.2307692 0.00000000 0.1111111 0.00000000
#>   [8,] 0.5714286 1.00000000 0.5714286 1.00000000
#>   [9,] 1.0000000 1.00000000 0.8181818 1.00000000
#>  [10,] 0.7142857 0.83333333 0.7777778 0.87500000
#>  [11,] 0.6666667 0.80000000 0.8000000 0.87500000
#>  [12,] 0.1176471 0.22222222 0.0000000 0.60000000
#>  [13,] 0.1428571 0.28571429 0.6428571 0.50000000
#>  [14,] 0.4000000 0.00000000 1.0000000 1.00000000
#>  [15,] 0.5555556 0.27272727 0.8000000 0.71428571
#>  [16,] 0.3333333 0.25000000 1.0000000 0.90909091
#>  [17,] 0.3333333 0.42857143 1.0000000 1.00000000
#>  [18,] 0.0000000 1.00000000 0.5000000 0.00000000
#>  [19,] 0.0000000 0.08333333 0.7142857 0.45454545
#>  [20,] 0.6666667 0.13333333 1.0000000 0.57142857
#>  [21,] 0.4000000 1.00000000 0.8888889 0.57142857
#>  [22,] 1.0000000 0.86666667 0.8333333 1.00000000
#>  [23,] 0.6666667 0.62500000 0.9375000 0.75000000
#>  [24,]        NA 0.00000000 1.0000000 0.00000000
#>  [25,] 0.0000000 0.00000000 0.7000000 0.62500000
#>  [26,] 1.0000000 0.16666667 1.0000000 0.85714286
#>  [27,] 0.9230769 0.36363636 0.7500000 0.90909091
#>  [28,] 0.6666667 0.12500000 0.8571429 0.87500000
#>  [29,] 0.7500000 0.25000000 0.4285714 0.50000000
#>  [30,] 0.0000000 0.14285714 1.0000000 0.50000000
#>  [31,] 1.0000000 0.28571429 1.0000000 1.00000000
#>  [32,] 0.4000000 1.00000000 0.6000000 1.00000000
#>  [33,] 0.6666667 0.50000000 1.0000000 1.00000000
#>  [34,] 0.1818182 0.37500000 0.7222222 0.76923077
#>  [35,] 0.9090909 0.75000000 1.0000000 0.75000000
#>  [36,] 0.8333333 1.00000000 0.7272727 0.69230769
#>  [37,]        NA 0.00000000 1.0000000 0.75000000
#>  [38,] 1.0000000 0.55555556 1.0000000 1.00000000
#>  [39,] 0.8750000 0.90000000 1.0000000 1.00000000
#>  [40,] 0.8888889 0.75000000 1.0000000 1.00000000
#>  [41,] 1.0000000 1.00000000 0.6000000 1.00000000
#>  [42,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [43,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [44,] 1.0000000 0.77777778 1.0000000 1.00000000
#>  [45,] 1.0000000 1.00000000 0.3333333 1.00000000
#>  [46,] 0.8333333 0.45454545 0.5000000 0.75000000
#>  [47,] 1.0000000 0.75000000 0.8333333 1.00000000
#>  [48,] 1.0000000 1.00000000 0.8571429 0.88888889
#>  [49,] 1.0000000 1.00000000 0.8571429 1.00000000
#>  [50,] 1.0000000 0.87500000 1.0000000 1.00000000
#>  [51,] 1.0000000 0.66666667 0.9090909 1.00000000
#>  [52,] 0.6000000 0.00000000 0.5714286 0.75000000
#>  [53,] 0.8888889 0.72727273 0.2500000 1.00000000
#>  [54,] 1.0000000 0.60000000 1.0000000 1.00000000
#>  [55,] 0.6666667 0.71428571 1.0000000 0.75000000
#>  [56,] 1.0000000 0.77777778 1.0000000 0.60000000
#>  [57,] 0.0000000 0.00000000 0.0000000 0.00000000
#>  [58,] 0.7142857 0.28571429 0.6666667 1.00000000
#>  [59,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [60,] 1.0000000 0.60000000 1.0000000 0.71428571
#>  [61,] 1.0000000 1.00000000 1.0000000 0.77777778
#>  [62,] 0.6363636 0.40000000 0.8000000 0.80000000
#>  [63,] 0.8750000 0.33333333 1.0000000 1.00000000
#>  [64,] 1.0000000 1.00000000 0.6666667 0.88888889
#>  [65,] 1.0000000 0.50000000 1.0000000 1.00000000
#>  [66,] 0.8000000 0.25000000 0.7142857 0.83333333
#>  [67,] 0.8888889 0.71428571 0.8888889 1.00000000
#>  [68,] 0.7777778 0.50000000 0.9090909 1.00000000
#>  [69,] 0.2000000 0.25000000 1.0000000 0.50000000
#>  [70,] 1.0000000 0.60000000 1.0000000 0.87500000
#>  [71,] 0.2857143 0.20000000 0.4285714 0.57142857
#>  [72,] 0.0000000 0.33333333 0.5000000 1.00000000
#>  [73,]        NA 0.00000000        NA 0.75000000
#>  [74,] 0.0000000 0.00000000 1.0000000 1.00000000
#>  [75,] 0.2000000 0.00000000 1.0000000 1.00000000
#>  [76,] 0.2000000 0.16666667 1.0000000 1.00000000
#>  [77,] 0.2857143 0.28571429 1.0000000 1.00000000
#>  [78,] 0.5000000 0.62500000 1.0000000 1.00000000
#>  [79,] 0.7500000 0.70000000 0.3333333 0.75000000
#>  [80,] 0.0000000 0.40000000 0.0000000 0.75000000
#>  [81,] 0.0000000 0.37500000 0.0000000 0.83333333
#>  [82,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [83,] 1.0000000 1.00000000 0.8333333 1.00000000
#>  [84,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [85,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [86,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [87,] 1.0000000 0.72727273 1.0000000 1.00000000
#>  [88,] 1.0000000 0.85714286 0.8000000 1.00000000
#>  [89,] 0.6666667 0.60000000 0.6000000         NA
#>  [90,] 1.0000000 0.25000000 1.0000000 1.00000000
#>  [91,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [92,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [93,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [94,] 1.0000000 0.00000000 0.3333333         NA
#>  [95,] 1.0000000 1.00000000 0.5000000 0.50000000
#>  [96,] 1.0000000 1.00000000 1.0000000 0.75000000
#>  [97,] 1.0000000 1.00000000 1.0000000 0.80000000
#>  [98,] 1.0000000 1.00000000 1.0000000 1.00000000
#>  [99,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [100,] 0.5000000 1.00000000 1.0000000 0.66666667
#> [101,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [102,] 1.0000000         NA 1.0000000         NA
#> [103,] 1.0000000 0.33333333 1.0000000 1.00000000
#> [104,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [105,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [106,] 1.0000000 0.75000000 1.0000000 1.00000000
#> [107,] 1.0000000 1.00000000 1.0000000 0.75000000
#> [108,] 0.7500000 1.00000000 1.0000000 1.00000000
#> [109,] 1.0000000 0.87500000 1.0000000         NA
#> [110,]        NA 1.00000000 1.0000000 0.66666667
#> [111,]        NA 1.00000000 1.0000000 1.00000000
#> [112,]        NA 1.00000000 1.0000000 1.00000000
#> [113,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [114,] 1.0000000 1.00000000 0.6666667 0.50000000
#> [115,] 0.5000000 0.50000000 1.0000000 1.00000000
#> [116,] 1.0000000 1.00000000 1.0000000 0.50000000
#> [117,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [118,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [119,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [120,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [121,] 1.0000000 0.75000000 1.0000000 1.00000000
#> [122,] 0.3750000 0.00000000 0.2000000 0.50000000
#> [123,] 1.0000000 1.00000000 0.8571429 0.60000000
#> [124,] 0.4000000 0.00000000 0.6666667 0.50000000
#> [125,] 0.8333333 0.50000000 0.0000000 0.50000000
#> [126,] 1.0000000 0.50000000 1.0000000 1.00000000
#> [127,] 0.8750000 0.50000000 0.8571429 1.00000000
#> [128,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [129,] 1.0000000 1.00000000 1.0000000 0.83333333
#> [130,] 1.0000000 0.80000000 1.0000000 0.91666667
#> [131,] 0.5000000 0.33333333 1.0000000 1.00000000
#> [132,] 1.0000000 1.00000000 0.2500000 1.00000000
#> [133,] 0.5000000 1.00000000 1.0000000 0.66666667
#> [134,] 0.7500000 0.70000000 1.0000000 0.85714286
#> [135,] 1.0000000 0.61538462 0.7500000 1.00000000
#> [136,] 1.0000000 0.85714286 1.0000000 0.85714286
#> [137,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [138,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [139,] 1.0000000 1.00000000 1.0000000 0.60000000
#> [140,] 1.0000000 1.00000000 1.0000000 0.80000000
#> [141,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [142,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [143,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [144,] 1.0000000 0.75000000 1.0000000 0.75000000
#> [145,] 0.8333333 0.83333333 1.0000000 0.75000000
#> [146,] 1.0000000 1.00000000 0.6666667 0.66666667
#> [147,] 1.0000000 1.00000000 0.7500000 1.00000000
#> [148,] 1.0000000 0.80000000 1.0000000 0.83333333
#> [149,] 0.4000000 0.85714286 0.7500000 0.81818182
#> [150,] 1.0000000 1.00000000 1.0000000 0.85714286
#> [151,] 0.8571429 0.90000000 1.0000000 1.00000000
#> [152,] 0.5000000 1.00000000 0.5000000 1.00000000
#> [153,] 0.6666667 1.00000000 1.0000000 1.00000000
#> [154,] 0.5000000 0.60000000 1.0000000 0.66666667
#> [155,] 0.8333333 0.50000000 0.5000000 0.60000000
#> [156,] 1.0000000 1.00000000 0.8181818 0.87500000
#> [157,] 1.0000000 1.00000000 0.8000000 0.85714286
#> [158,] 1.0000000 0.87500000 0.7857143 1.00000000
#> [159,] 1.0000000 1.00000000 0.7777778 1.00000000
#> [160,] 1.0000000 0.83333333 0.8461538 1.00000000
#> [161,] 1.0000000 0.80000000 0.7500000 0.88888889
#> [162,] 1.0000000 0.60000000 0.7500000 0.87500000
#> [163,] 0.7500000 0.75000000 0.8181818 0.87500000
#> [164,] 0.6666667 0.66666667 0.7500000 0.66666667
#> [165,] 1.0000000 0.20000000 1.0000000 1.00000000
#> [166,] 1.0000000 0.50000000 1.0000000 1.00000000
#> [167,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [168,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [169,] 0.6666667 0.00000000 0.2500000 0.40000000
#> [170,] 0.6666667 1.00000000 0.7500000 1.00000000
#> [171,] 0.8571429 1.00000000 0.8333333 0.00000000
#> [172,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [173,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [174,] 1.0000000 0.42857143 0.7500000 1.00000000
#> [175,] 0.6000000 0.50000000 0.7142857 1.00000000
#> [176,] 1.0000000 0.00000000 0.8750000 1.00000000
#> [177,] 1.0000000 0.71428571 1.0000000 1.00000000
#> [178,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [179,] 1.0000000 0.90909091 1.0000000 0.75000000
#> [180,] 1.0000000 0.75000000 0.6666667 0.75000000
#> [181,] 1.0000000 0.77777778 1.0000000 0.66666667
#> [182,]        NA 0.60000000 1.0000000 0.50000000
#> [183,] 0.7500000 0.00000000 1.0000000 0.85714286
#> [184,] 0.5000000 0.00000000 0.0000000 0.54545455
#> [185,] 0.4000000 0.90000000 0.2000000 0.83333333
#> [186,] 0.2500000 0.88888889 0.3333333 0.50000000
#> [187,] 0.0000000 0.00000000 0.0000000 0.20000000
#> [188,] 0.6666667 0.75000000 0.2000000 0.50000000
#> [189,] 0.5000000 1.00000000 0.5000000 0.83333333
#> [190,] 0.3333333 0.83333333 0.5000000 0.60000000
#> [191,] 0.3333333 0.75000000 0.7500000 0.66666667
#> [192,] 1.0000000         NA 0.7500000 0.66666667
#> [193,] 1.0000000 1.00000000 0.7500000 1.00000000
#> [194,] 0.6666667 1.00000000 0.6666667 0.33333333
#> [195,] 0.0000000 0.50000000 0.5714286 0.40000000
#> [196,] 0.0000000 0.00000000 0.0000000         NA
#> [197,] 0.6666667 0.00000000 0.2727273 0.80000000
#> [198,]        NA 0.50000000 0.0000000 0.50000000
#> [199,]        NA 0.66666667 0.2000000 0.50000000
#> [200,] 1.0000000 0.75000000 0.4545455 0.50000000
#> [201,] 0.0000000 0.80000000 0.4545455 0.50000000
#> [202,] 0.1428571 0.60000000 0.3750000 0.00000000
#> [203,] 0.0000000 0.22222222 0.0000000 0.00000000
#> [204,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [205,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [206,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [207,]        NA 0.20000000        NA 1.00000000
#> [208,] 0.0000000 0.00000000        NA 0.00000000
#> [209,]        NA 0.00000000        NA 0.00000000
#> [210,] 0.2000000 0.00000000 0.2500000 0.00000000
#> [211,] 0.3333333 0.00000000 0.0000000 0.33333333
#> [212,] 0.1428571 0.00000000 0.0000000 0.00000000
#> [213,] 0.3333333 0.00000000 0.0000000 0.00000000
#> [214,] 0.3333333 0.00000000 0.0000000 0.00000000
#> [215,] 0.2000000 0.00000000 0.0000000 0.00000000
#> [216,] 0.2500000         NA 0.0000000 0.00000000
#> [217,] 1.0000000 0.00000000 0.0000000 0.00000000
#> [218,] 1.0000000 0.00000000 0.0000000 0.00000000
#> [219,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [220,] 0.2500000 0.50000000 0.0000000 0.14285714
#> [221,] 0.2500000 0.33333333 0.0000000 0.00000000
#> [222,] 0.2500000 0.00000000 0.0000000 0.00000000
#> [223,] 0.2500000 0.50000000 0.0000000 0.08333333
#> [224,] 0.2500000 0.50000000 0.3333333 0.00000000
#> [225,] 1.0000000 0.33333333 1.0000000 0.08333333
#> [226,] 0.5000000 0.33333333 0.0000000 0.00000000
#> [227,] 0.5000000 0.50000000 0.0000000 0.00000000
#> [228,] 0.0000000 1.00000000 0.0000000 0.00000000
#> [229,] 0.5000000 0.00000000 0.0000000 0.00000000
#> [230,] 0.6666667 0.00000000 0.0000000 0.00000000
#> [231,] 0.0000000 0.00000000        NA         NA
#> [232,] 0.3333333 0.00000000 0.0000000 0.00000000
#> [233,] 0.5000000 0.00000000 0.0000000 0.00000000
#> [234,]        NA 0.00000000        NA 0.00000000
#> [235,]        NA 0.00000000 0.0000000         NA
#> [236,]        NA 0.00000000 0.0000000 0.00000000
#> [237,]        NA         NA 0.0000000 0.00000000
#> [238,]        NA         NA 0.0000000         NA
#> [239,]        NA 0.00000000 0.0000000 0.00000000
#> [240,]        NA 0.00000000 0.0000000 0.00000000
#> [241,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [242,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [243,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [244,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [245,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [246,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [247,] 0.5000000 0.00000000 0.0000000 0.00000000
#> [248,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [249,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [250,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [251,] 0.0000000         NA 0.0000000 0.00000000
#> [252,] 0.0000000         NA        NA 0.00000000
#> [253,] 0.0000000         NA        NA 0.00000000
#> [254,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [255,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [256,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [257,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [258,] 0.2000000 0.00000000 0.0000000 0.00000000
#> [259,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [260,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [261,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [262,] 0.1428571         NA 0.0000000 0.00000000
#> [263,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [264,]        NA 0.50000000 0.0000000 0.00000000
#> [265,]        NA 0.00000000 0.0000000 0.00000000
#> [266,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [267,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [268,] 0.0000000 0.33333333 0.0000000 0.00000000
#> [269,] 0.0000000 0.33333333 0.0000000 0.00000000
#> [270,] 0.0000000         NA 0.0000000 0.33333333
#> [271,] 0.0000000         NA 0.0000000 0.00000000
#> [272,] 0.0000000         NA 0.0000000 0.00000000
#> [273,] 0.0000000         NA 0.0000000 0.00000000
#> [274,] 0.0000000         NA 0.0000000 0.00000000
#> [275,] 0.0000000         NA 0.0000000 0.00000000
#> [276,]        NA         NA 0.0000000 0.00000000
#> [277,]        NA         NA 0.0000000 0.00000000
#> [278,]        NA 0.00000000 0.0000000         NA
#> [279,] 0.0000000 0.50000000 0.0000000         NA
#> [280,] 0.0000000 0.20000000        NA 0.00000000
#> [281,] 0.0000000 0.00000000        NA 0.00000000
#> [282,] 0.2500000 0.40000000 0.0000000 0.00000000
#> [283,] 0.2000000 0.16666667 0.0000000 0.00000000
#> [284,] 0.4000000 0.00000000 0.0000000 0.00000000
#> [285,] 0.2500000 0.00000000 0.0000000 0.00000000
#> [286,] 0.2000000 0.00000000 0.0000000 0.00000000
#> [287,] 0.0000000 1.00000000 0.0000000 0.00000000
#> [288,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [289,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [290,] 1.0000000         NA 0.0000000 0.00000000
#> [291,] 1.0000000 0.00000000 0.0000000 0.00000000
#> [292,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [293,] 1.0000000 1.00000000 0.0000000 0.00000000
#> [294,] 0.5000000 0.00000000 0.0000000 0.00000000
#> [295,] 0.5000000 0.00000000 0.0000000 0.00000000
#> [296,] 0.0000000 0.33333333        NA 0.00000000
#> [297,] 0.0000000 0.00000000        NA         NA
#> [298,]        NA 0.00000000        NA 0.00000000
#> [299,] 1.0000000         NA 0.0000000 0.00000000
#> [300,] 0.0000000         NA 0.0000000 0.00000000
#> [301,] 0.5000000 0.00000000 0.0000000 0.00000000
#> [302,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [303,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [304,] 0.0000000 0.00000000 0.0000000 0.50000000
#> [305,] 0.5000000 0.00000000 0.0000000         NA
#> [306,] 0.6666667 0.25000000 0.0000000 0.00000000
#> [307,] 0.6666667 0.25000000 0.0000000 0.00000000
#> [308,] 1.0000000 0.66666667 0.0000000 0.00000000
#> [309,] 1.0000000 0.66666667 0.0000000 0.00000000
#> [310,] 1.0000000 0.66666667 0.0000000 0.00000000
#> [311,] 1.0000000 1.00000000 0.0000000 0.00000000
#> [312,]        NA 1.00000000 0.0000000 0.00000000
#> [313,]        NA 0.50000000 0.0000000 0.33333333
#> [314,]        NA         NA 0.0000000 0.00000000
#> [315,]        NA         NA 0.0000000 0.00000000
#> [316,]        NA         NA 0.0000000 0.00000000
#> [317,] 0.0000000         NA 0.0000000 0.16666667
#> [318,] 0.0000000 1.00000000 0.0000000 0.00000000
#> [319,] 0.0000000 1.00000000 0.0000000 0.00000000
#> [320,] 0.4000000 1.00000000 0.0000000 0.00000000
#> [321,] 0.5000000 1.00000000 1.0000000 0.00000000
#> [322,] 0.6666667         NA 0.0000000 0.00000000
#> [323,] 0.5000000         NA        NA 0.00000000
#> [324,] 0.0000000         NA        NA 0.00000000
#> [325,] 1.0000000         NA        NA 0.00000000
#> [326,] 1.0000000         NA        NA 0.00000000
#> [327,] 1.0000000         NA 0.0000000 0.00000000
#> [328,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [329,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [330,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [331,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [332,] 0.0000000 0.00000000 0.5000000 0.00000000
#> [333,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [334,] 1.0000000 0.00000000 0.0000000 0.00000000
#> [335,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [336,]        NA 0.00000000 0.5000000 0.00000000
#> [337,] 1.0000000 0.50000000 0.0000000 0.25000000
#> [338,] 1.0000000 1.00000000 0.0000000 0.25000000
#> [339,] 1.0000000 1.00000000 0.2000000 0.12500000
#> [340,] 1.0000000 0.33333333 0.2500000 0.50000000
#> [341,] 1.0000000 0.33333333 0.0000000 0.75000000
#> [342,] 0.6666667 0.33333333 0.0000000 0.25000000
#> [343,] 0.5000000 0.00000000        NA 0.00000000
#> [344,] 0.7500000 0.00000000 0.0000000 0.00000000
#> [345,] 0.2500000 0.00000000 0.0000000 0.00000000
#> [346,] 0.3333333 0.00000000 0.0000000 0.00000000
#> [347,] 0.3333333 0.50000000 0.0000000 0.00000000
#> [348,] 0.3333333 0.25000000 0.0000000 0.00000000
#> [349,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [350,] 0.6666667 1.00000000 0.0000000 0.00000000
#> [351,] 0.6666667 1.00000000 0.0000000 0.00000000
#> [352,] 0.5000000 0.60000000 0.0000000 0.00000000
#> [353,] 0.3333333 0.83333333 0.0000000 0.00000000
#> [354,] 0.8181818 0.85714286 0.0000000 0.00000000
#> [355,] 0.9000000 1.00000000 0.0000000 0.00000000
#> [356,] 0.8888889 1.00000000 0.0000000 0.00000000
#> [357,] 0.5454545 0.50000000 0.1111111 0.00000000
#> [358,] 0.5454545 0.66666667 0.1666667 0.00000000
#> [359,] 0.7500000 0.33333333 0.3333333 0.33333333
#> [360,] 0.3333333 0.66666667 0.1428571 0.00000000
#> [361,] 0.3333333 0.50000000 0.1428571 0.00000000
#> [362,] 1.0000000 0.66666667 0.1250000 0.60000000
#> [363,] 0.8333333 0.66666667 0.0000000 0.25000000
#> [364,] 0.7500000 0.55555556 0.0000000 0.25000000
#> [365,] 0.7500000 0.50000000 0.0000000 0.25000000
#> [366,] 0.6666667 0.11111111 0.0000000 0.00000000
#> [367,]        NA 1.00000000 0.0000000 0.00000000
#> [368,]        NA 1.00000000 0.0000000 0.25000000
#> [369,]        NA 1.00000000 0.3333333 0.20000000
#> [370,]        NA 1.00000000 0.0000000 0.20000000
#> [371,] 0.5000000 0.50000000 0.4285714 0.12500000
#> [372,]        NA 1.00000000 0.3333333 0.00000000
#> [373,]        NA 0.50000000 0.6666667 0.00000000
#> [374,]        NA         NA 0.6666667 0.00000000
#> [375,]        NA         NA        NA 1.00000000
#> [376,]        NA         NA        NA 1.00000000
#> [377,]        NA         NA 0.0000000         NA
#> [378,]        NA         NA        NA 0.00000000
#> [379,] 0.0000000 0.33333333 0.2000000 0.00000000
#> [380,] 0.0000000 0.14285714 0.4000000 0.20000000
#> [381,] 0.0000000 0.30769231 0.2000000 0.00000000
#> [382,] 0.0000000 0.18750000 0.0000000 0.08333333
#> [383,] 0.0000000 0.73333333 0.1666667 0.22222222
#> [384,] 0.1428571 0.83333333 0.4285714 0.20000000
#> [385,] 0.1666667 0.90000000 0.2500000 0.00000000
#> [386,] 0.4285714 1.00000000 0.3333333 0.00000000
#> [387,] 0.0000000 0.42857143 0.6000000 0.00000000
#> [388,] 0.1666667 0.50000000 0.5000000 0.07692308
#> [389,] 0.1428571 0.28571429 0.0000000 0.00000000
#> [390,] 0.1000000 0.14285714 0.4444444 0.14285714
#> [391,] 0.0000000 0.66666667 0.2000000 0.33333333
#> [392,] 0.2500000 0.66666667 0.2000000 0.12500000
#> [393,] 0.3333333 0.00000000 0.2222222 0.75000000
#> [394,] 0.2500000 0.00000000 0.2307692 1.00000000
#> [395,] 0.0000000 0.14285714 0.2000000 0.00000000
#> [396,] 0.0000000 0.14285714 0.4000000 0.00000000
#> [397,] 0.0000000 0.11111111 0.4615385 0.00000000
#> [398,] 1.0000000 0.66666667 1.0000000 1.00000000
#> [399,] 0.7500000 0.66666667 0.8750000 0.55555556
#> [400,] 0.4000000 1.00000000 1.0000000 0.50000000
#> [401,] 1.0000000 1.00000000 1.0000000 0.40000000
#> [402,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [403,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [404,] 0.8571429 0.83333333 0.8888889 1.00000000
#> [405,] 1.0000000 1.00000000 0.0000000 0.66666667
#> [406,] 0.0000000 0.11111111 0.5000000 0.66666667
#> [407,] 0.0000000 0.50000000 0.7500000 0.33333333
#> [408,] 0.0000000 0.27272727 0.2000000 0.41666667
#> [409,] 0.4000000 0.00000000 0.2500000 0.75000000
#> [410,] 0.5000000 0.50000000 0.8333333 1.00000000
#> [411,] 0.0000000 1.00000000 0.0000000 0.77777778
#> [412,] 1.0000000 0.75000000 0.8571429 1.00000000
#> [413,] 1.0000000 0.60000000 1.0000000 0.75000000
#> [414,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [415,] 1.0000000 1.00000000 0.0000000 1.00000000
#> [416,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [417,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [418,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [419,] 0.0000000 0.33333333 0.5000000 0.66666667
#> [420,] 0.7500000 0.22222222 0.9166667 0.25000000
#> [421,] 1.0000000 0.27272727 0.7777778 0.90000000
#> [422,] 0.2500000 0.54545455 0.5714286 1.00000000
#> [423,] 0.0000000 0.66666667 0.5000000 1.00000000
#> [424,] 0.0000000 0.37500000 0.0000000 0.00000000
#> [425,] 0.2000000 0.28571429 0.4285714 0.00000000
#> [426,] 0.2500000 0.11111111 0.3636364 0.20000000
#> [427,] 0.0000000 0.10000000 0.5000000 0.63636364
#> [428,] 0.5000000 0.46153846 0.7857143 0.62500000
#> [429,] 0.8000000 0.30000000 0.8750000 0.87500000
#> [430,] 0.8000000 0.00000000 0.6666667 1.00000000
#> [431,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [432,] 1.0000000 1.00000000 1.0000000 0.80000000
#> [433,] 0.7500000 0.88888889 1.0000000 1.00000000
#> [434,] 0.0000000 0.66666667        NA 0.60000000
#> [435,] 0.6000000 0.25000000 0.6000000 0.75000000
#> [436,] 1.0000000 0.66666667 1.0000000 1.00000000
#> [437,] 0.0000000 0.00000000 1.0000000 0.90909091
#> [438,] 0.5000000 0.33333333 1.0000000 1.00000000
#> [439,] 0.7500000 0.25000000 1.0000000 0.66666667
#> [440,] 0.2500000 0.66666667 0.5000000 0.75000000
#> [441,] 0.6000000 0.33333333 0.0000000 0.80000000
#> [442,] 1.0000000 0.16666667 0.7500000 1.00000000
#> [443,] 1.0000000 0.20000000 1.0000000 1.00000000
#> [444,] 0.5000000 0.81818182 1.0000000 0.63636364
#> [445,] 1.0000000 0.88888889 1.0000000 1.00000000
#> [446,] 0.5000000 0.36363636 1.0000000 0.62500000
#> [447,] 0.8000000 0.44444444 1.0000000 1.00000000
#> [448,] 0.6666667 0.09090909 1.0000000 1.00000000
#> [449,] 0.2500000 0.20000000 0.6666667 0.66666667
#> [450,] 0.1428571 0.20000000 0.8000000 0.60000000
#> [451,] 0.5000000 0.77777778 1.0000000 1.00000000
#> [452,] 1.0000000 1.00000000 1.0000000 0.75000000
#> [453,] 1.0000000 0.87500000 1.0000000 1.00000000
#> [454,]        NA 0.75000000 0.7500000 0.66666667
#> [455,]        NA 1.00000000 0.6666667 0.80000000
#> [456,] 0.5000000 0.75000000 0.6250000 0.83333333
#> [457,] 1.0000000 0.44444444 0.8571429 0.81818182
#> [458,] 1.0000000 1.00000000 0.9166667 1.00000000
#> [459,] 0.7272727 0.20000000 0.8000000 0.80000000
#> [460,] 0.1666667 0.16666667 0.2500000 0.25000000
#> [461,] 0.5000000 0.25000000 1.0000000 0.75000000
#> [462,] 0.8000000 0.00000000 1.0000000 0.75000000
#> [463,] 1.0000000 1.00000000 0.6666667 1.00000000
#> [464,] 0.0000000 0.00000000 0.0000000 0.00000000
#> [465,] 1.0000000 0.50000000 0.0000000         NA
#> [466,] 0.0000000 0.33333333 0.0000000 0.00000000
#> [467,] 0.0000000 0.00000000 0.0000000 1.00000000
#> [468,] 0.0000000 0.25000000 0.2857143 0.00000000
#> [469,] 0.3333333 0.00000000 1.0000000 1.00000000
#> [470,] 1.0000000 0.80000000 1.0000000         NA
#> [471,] 1.0000000 0.83333333 1.0000000         NA
#> [472,] 1.0000000 0.16666667 1.0000000 0.50000000
#> [473,] 0.1250000 0.22222222 0.2857143 0.66666667
#> [474,] 0.1428571 0.30000000 0.3750000 0.66666667
#> [475,] 0.3000000 0.00000000 0.4285714 1.00000000
#> [476,] 0.1000000 0.00000000 0.5000000 0.90000000
#> [477,] 0.2727273 0.20000000 0.6666667 0.60000000
#> [478,] 0.3333333 0.00000000 1.0000000 1.00000000
#> [479,] 0.5000000 0.00000000 1.0000000 1.00000000
#> [480,] 1.0000000 0.00000000 0.6666667 1.00000000
#> [481,] 0.2500000 0.00000000 0.2500000 0.50000000
#> [482,] 1.0000000 0.33333333 1.0000000 0.87500000
#> [483,] 1.0000000 0.25000000 1.0000000 1.00000000
#> [484,] 0.9285714 0.50000000 0.7777778 1.00000000
#> [485,] 1.0000000 0.28571429 0.7142857 0.88888889
#> [486,] 1.0000000 0.50000000 1.0000000 0.85714286
#> [487,] 0.1428571 0.25000000 0.8571429 0.75000000
#> [488,] 0.9090909 0.00000000 1.0000000 0.90000000
#> [489,] 0.8571429 0.80000000 0.7142857 1.00000000
#> [490,] 1.0000000 0.33333333 0.7500000 0.66666667
#> [491,] 0.2500000 0.00000000 0.5000000 0.00000000
#> [492,] 1.0000000 0.81818182 0.9166667 1.00000000
#> [493,] 1.0000000 0.90000000 0.7777778 1.00000000
#> [494,] 1.0000000 0.09090909 0.5714286 0.85714286
#> [495,] 0.7142857 0.62500000 1.0000000 1.00000000
#> [496,] 0.8888889 0.66666667 1.0000000 1.00000000
#> [497,] 0.7500000 0.85714286 0.7500000 1.00000000
#> [498,] 0.8750000 0.60000000 1.0000000 0.88888889
#> [499,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [500,] 1.0000000 0.00000000 0.7500000 1.00000000
#> [501,] 0.4000000 0.00000000 1.0000000 0.50000000
#> [502,] 0.1428571 0.16666667 1.0000000 0.40000000
#> [503,] 0.4285714 0.42857143 0.8181818 1.00000000
#> [504,] 0.5714286 0.75000000 1.0000000 1.00000000
#> [505,] 0.0000000 0.00000000 0.5000000 0.50000000
#> [506,] 0.7272727 0.66666667 0.8333333 0.57142857
#> [507,] 0.7500000 0.71428571 0.8333333 0.88888889
#> [508,] 0.6363636 1.00000000 0.8000000 0.83333333
#> [509,] 0.7500000 0.85714286 0.8750000 0.60000000
#> [510,] 0.4285714 0.00000000 0.7142857 1.00000000
#> [511,] 1.0000000 0.83333333 0.8571429 1.00000000
#> [512,] 0.1666667 0.50000000 0.5000000 1.00000000
#> [513,] 0.3333333 0.66666667 1.0000000 0.80000000
#> [514,] 0.6666667 0.80000000 1.0000000 1.00000000
#> [515,] 1.0000000 1.00000000 0.5000000 1.00000000
#> [516,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [517,] 1.0000000 1.00000000 0.8571429 1.00000000
#> [518,] 1.0000000 0.00000000 1.0000000 1.00000000
#> [519,] 0.0000000 0.50000000 1.0000000 0.75000000
#> [520,] 0.0000000 0.00000000        NA 0.50000000
#> [521,] 0.8333333 0.22222222 0.9090909 0.92307692
#> [522,] 0.0000000 0.25000000 0.0000000 0.57142857
#> [523,] 0.0000000 0.00000000        NA 0.66666667
#> [524,] 0.4285714 0.40000000 0.8888889 0.50000000
#> [525,] 0.6250000 0.30000000 0.5000000 0.16666667
#> [526,] 0.0000000 0.00000000 0.6666667 0.57142857
#> [527,] 0.0000000 0.00000000 0.0000000 0.06250000
#> [528,] 0.2857143 0.25000000 0.6666667 0.80000000
#> [529,] 0.0000000 0.00000000 0.4000000 1.00000000
#> [530,] 0.8888889 0.42857143 0.6666667 1.00000000
#> [531,] 1.0000000 0.66666667 0.0000000 1.00000000
#> [532,] 0.3333333 0.50000000 1.0000000 0.33333333
#> [533,] 0.6666667 1.00000000 0.6666667 1.00000000
#> [534,] 0.7500000 0.50000000 0.5000000 1.00000000
#> [535,] 0.0000000 1.00000000 0.0000000         NA
#> [536,] 0.0000000 0.00000000 0.0000000         NA
#> [537,] 1.0000000 0.16666667 0.6000000 1.00000000
#> [538,] 0.6000000 0.00000000 0.6666667 0.00000000
#> [539,] 0.5714286 0.75000000 0.8571429 1.00000000
#> [540,] 0.5714286 0.75000000 0.8571429 1.00000000
#> [541,]        NA 0.00000000        NA 1.00000000
#> [542,] 0.4285714 0.75000000 0.1250000 0.60000000
#> [543,] 0.3750000 0.63636364 0.7500000 0.50000000
#> [544,] 1.0000000 0.90909091 0.8000000 1.00000000
#> [545,] 1.0000000 0.92307692 1.0000000 1.00000000
#> [546,] 1.0000000 0.33333333 1.0000000 1.00000000
#> [547,] 1.0000000         NA        NA         NA
#> [548,] 0.0000000         NA        NA         NA
#> [549,] 0.0000000         NA        NA         NA
#> [550,]        NA         NA 0.0000000         NA
#> [551,]        NA         NA 1.0000000         NA
#> [552,]        NA         NA 1.0000000         NA
#> [553,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [554,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [555,] 1.0000000 1.00000000 1.0000000         NA
#> [556,]        NA 1.00000000        NA         NA
#> [557,] 0.0000000         NA 1.0000000 1.00000000
#> [558,] 1.0000000         NA 1.0000000 1.00000000
#> [559,] 1.0000000 0.80000000 1.0000000 1.00000000
#> [560,] 1.0000000 1.00000000 1.0000000 0.87500000
#> [561,] 1.0000000 1.00000000 0.8750000 1.00000000
#> [562,] 1.0000000 1.00000000 0.9166667 1.00000000
#> [563,] 0.6666667 0.00000000 1.0000000         NA
#> [564,]        NA 1.00000000        NA         NA
#> [565,]        NA 1.00000000        NA         NA
#> [566,] 1.0000000 1.00000000        NA         NA
#> [567,] 0.5714286 0.50000000 0.0000000 1.00000000
#> [568,] 0.5000000 0.22222222 0.5000000 1.00000000
#> [569,] 0.6666667 0.25000000 0.4000000 0.75000000
#> [570,] 1.0000000 0.18181818 1.0000000 1.00000000
#> [571,] 0.7272727 0.11111111 0.9166667 0.37500000
#> [572,] 0.2000000 0.33333333 1.0000000 1.00000000
#> [573,] 0.1250000 0.33333333 0.9285714         NA
#> [574,] 0.0000000 0.20000000 0.8000000 1.00000000
#> [575,] 0.0000000 0.31250000 0.8333333 1.00000000
#> [576,] 0.6666667 0.41666667 0.6666667 0.90909091
#> [577,] 0.5000000 0.12500000 0.5333333 0.88888889
#> [578,] 0.7142857 0.87500000 1.0000000 0.75000000
#> [579,] 1.0000000 0.90909091 1.0000000 1.00000000
#> [580,]        NA 0.00000000        NA         NA
#> [581,]        NA 1.00000000        NA         NA
#> [582,] 1.0000000         NA        NA 0.00000000
#> [583,] 1.0000000         NA        NA         NA
#> [584,]        NA 0.00000000        NA         NA
#> [585,] 0.5000000 1.00000000 0.0000000 0.50000000
#> [586,] 1.0000000 0.00000000 0.2500000 0.42857143
#> [587,] 1.0000000 0.75000000 0.5000000 1.00000000
#> [588,]        NA 1.00000000 1.0000000 1.00000000
#> [589,] 0.6000000 0.60000000 0.5000000 0.60000000
#> [590,] 0.6000000 0.16666667 0.5000000 0.53846154
#> [591,] 0.8000000 0.25000000 0.5000000 0.80000000
#> [592,] 0.5000000 0.66666667 0.7777778 1.00000000
#> [593,] 0.1428571 0.33333333 0.5833333 1.00000000
#> [594,] 1.0000000 0.83333333 0.8000000 0.57142857
#> [595,] 1.0000000 0.66666667 0.8000000 1.00000000
#> [596,]        NA 1.00000000 0.8888889 1.00000000
#> [597,]        NA 1.00000000 0.8888889 0.87500000
#> [598,] 1.0000000 0.60000000 1.0000000 1.00000000
#> [599,] 0.2500000 0.00000000 0.8000000 0.66666667
#> [600,] 1.0000000 1.00000000 0.7142857 0.80000000
#> [601,] 1.0000000 0.60000000 0.6666667 0.80000000
#> [602,] 1.0000000 0.45454545 1.0000000 1.00000000
#> [603,] 1.0000000 0.54545455 1.0000000 1.00000000
#> [604,] 1.0000000 0.66666667 0.7500000 1.00000000
#> [605,]        NA 0.50000000 0.6666667 0.33333333
#> [606,] 0.5000000 0.80000000 0.8000000 0.25000000
#> [607,] 1.0000000 0.33333333 1.0000000 0.66666667
#> [608,] 0.6000000 1.00000000 1.0000000 0.87500000
#> [609,]        NA 0.00000000 1.0000000 0.60000000
#> [610,] 0.2000000 1.00000000 1.0000000 0.33333333
#> [611,] 1.0000000 1.00000000 1.0000000 0.85714286
#> [612,] 1.0000000 0.50000000 0.8000000 0.60000000
#> [613,] 0.5000000 0.25000000 0.8000000 0.88888889
#> [614,] 1.0000000 1.00000000 0.8750000 1.00000000
#> [615,] 0.8750000 1.00000000 1.0000000 1.00000000
#> [616,] 0.6666667 0.80000000 1.0000000 1.00000000
#> [617,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [618,] 0.0000000 1.00000000 1.0000000 1.00000000
#> [619,] 1.0000000 1.00000000 1.0000000 0.87500000
#> [620,] 0.6666667 0.16666667 0.8571429 1.00000000
#> [621,] 1.0000000 0.00000000 0.2500000 0.50000000
#> [622,] 0.4000000 0.12500000 0.8333333 0.33333333
#> [623,] 1.0000000 0.20000000 0.6000000 0.50000000
#> [624,] 1.0000000 0.45454545 1.0000000 0.80000000
#> [625,] 0.0000000 0.16666667 0.2000000 0.33333333
#> [626,] 0.7500000 0.30000000 0.7142857 0.80000000
#> [627,] 1.0000000 0.42857143 1.0000000 0.83333333
#> [628,] 0.8750000 0.66666667 0.8571429 0.80000000
#> [629,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [630,] 1.0000000 0.60000000 1.0000000 1.00000000
#> [631,] 0.6666667 0.60000000 0.8333333 0.66666667
#> [632,] 1.0000000 1.00000000 0.9000000 1.00000000
#> [633,] 0.8333333 0.66666667 1.0000000 0.71428571
#> [634,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [635,] 0.6000000 0.00000000 0.6666667 0.75000000
#> [636,] 0.0000000 0.00000000 0.0000000 0.25000000
#> [637,] 0.5000000 0.00000000 1.0000000 0.63636364
#> [638,] 0.3333333 0.00000000 0.5000000 0.50000000
#> [639,]        NA 0.00000000 0.6666667 1.00000000
#> [640,] 1.0000000 0.66666667 0.6666667 0.71428571
#> [641,] 1.0000000 0.50000000 0.6666667 0.33333333
#> [642,] 0.6666667 0.66666667 0.8571429 0.83333333
#> [643,] 0.8000000 0.81818182 0.8750000 0.83333333
#> [644,] 0.8571429 1.00000000 1.0000000 1.00000000
#> [645,] 0.5000000 0.00000000 1.0000000 0.71428571
#> [646,] 0.6666667         NA 1.0000000 0.71428571
#> [647,] 0.5000000 1.00000000 1.0000000 0.85714286
#> [648,] 0.0000000 0.66666667 0.7500000 0.87500000
#> [649,] 0.5000000 0.22222222 1.0000000 1.00000000
#> [650,] 0.5000000 0.50000000 1.0000000 1.00000000
#> [651,] 0.6000000 0.33333333 0.6666667 1.00000000
#> [652,] 0.9000000 0.80000000 0.8333333 0.90909091
#> [653,] 1.0000000 1.00000000 0.9000000 1.00000000
#> [654,] 1.0000000 1.00000000 1.0000000 0.71428571
#> [655,] 0.2500000 1.00000000 1.0000000 0.90000000
#> [656,] 0.6666667 1.00000000 1.0000000 1.00000000
#> [657,] 0.2000000 0.37500000 0.8333333 0.83333333
#> [658,] 1.0000000 0.33333333 1.0000000 0.77777778
#> [659,] 0.7500000 1.00000000 1.0000000 1.00000000
#> [660,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [661,]        NA 1.00000000 1.0000000 0.75000000
#> [662,] 0.5714286 0.14285714 0.9000000 0.77777778
#> [663,] 0.5000000 0.00000000 0.5000000 0.66666667
#> [664,] 0.8000000 0.25000000 1.0000000 1.00000000
#> [665,] 0.7692308 0.12500000 1.0000000 0.85714286
#> [666,] 0.4615385 0.37500000 1.0000000 0.85714286
#> [667,] 0.8000000 0.85714286 0.8333333 0.85714286
#> [668,] 0.7500000 0.66666667 0.5714286 1.00000000
#> [669,]        NA 0.00000000 1.0000000 1.00000000
#> [670,] 0.4444444 0.66666667 1.0000000 1.00000000
#> [671,] 0.0000000 0.00000000 0.7500000 1.00000000
#> [672,] 0.0000000 0.00000000 0.8750000 0.88888889
#> [673,] 1.0000000 0.33333333 1.0000000 1.00000000
#> [674,] 0.6000000 0.20000000 1.0000000 1.00000000
#> [675,] 0.5000000 0.18181818 1.0000000 1.00000000
#> [676,] 0.4000000 0.14285714 1.0000000 0.80000000
#> [677,]        NA 0.75000000 0.5000000 1.00000000
#> [678,] 1.0000000 0.60000000 0.5000000 0.71428571
#> [679,] 1.0000000 0.33333333 0.6666667 1.00000000
#> [680,] 0.3333333 0.33333333 0.5000000 0.83333333
#> [681,] 0.5000000 0.25000000 1.0000000 1.00000000
#> [682,] 0.6666667 0.20000000 1.0000000 1.00000000
#> [683,] 0.6666667 1.00000000 1.0000000 1.00000000
#> [684,] 0.5000000 0.33333333 0.5000000 0.87500000
#> [685,] 1.0000000 1.00000000 0.7500000 1.00000000
#> [686,] 0.8888889 1.00000000 1.0000000 1.00000000
#> [687,] 0.2000000 0.25000000 0.6666667 1.00000000
#> [688,] 0.6000000 0.33333333 0.6666667 0.75000000
#> [689,] 0.7500000 0.50000000 0.8333333 0.87500000
#> [690,] 0.6000000 0.66666667 0.7000000 0.90000000
#> [691,] 0.6000000 0.00000000 1.0000000 0.80000000
#> [692,] 0.5000000 0.37500000 1.0000000 1.00000000
#> [693,] 1.0000000 0.50000000 1.0000000 1.00000000
#> [694,] 1.0000000         NA 1.0000000 1.00000000
#> [695,] 0.4285714 1.00000000 1.0000000 1.00000000
#> [696,] 0.6666667 0.50000000 1.0000000 1.00000000
#> [697,] 0.5000000 0.00000000 0.9000000 0.83333333
#> [698,] 1.0000000 0.28571429 1.0000000 1.00000000
#> [699,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [700,] 1.0000000 0.66666667 1.0000000 1.00000000
#> [701,] 1.0000000 1.00000000 0.5000000 1.00000000
#> [702,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [703,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [704,] 1.0000000 0.50000000 0.8888889 1.00000000
#> [705,] 1.0000000 1.00000000 0.8571429 1.00000000
#> [706,] 0.8333333 1.00000000 1.0000000 0.33333333
#> [707,] 0.7500000 0.25000000 0.5000000 1.00000000
#> [708,] 1.0000000         NA 0.5000000 1.00000000
#> [709,] 0.5000000         NA 1.0000000 1.00000000
#> [710,] 1.0000000 1.00000000 1.0000000         NA
#> [711,] 1.0000000 0.66666667 0.7500000         NA
#> [712,] 1.0000000 0.75000000 1.0000000         NA
#> [713,] 0.7500000 0.14285714 0.8888889 1.00000000
#> [714,] 0.8888889 1.00000000 0.6666667 1.00000000
#> [715,] 0.8461538 0.83333333 0.8000000 0.20000000
#> [716,] 0.5454545 0.00000000 1.0000000 0.50000000
#> [717,] 1.0000000 0.83333333 1.0000000 0.76923077
#> [718,] 1.0000000 0.00000000 1.0000000 1.00000000
#> [719,] 1.0000000 0.00000000 1.0000000 1.00000000
#> [720,] 1.0000000 1.00000000 1.0000000 0.62500000
#> [721,] 0.8333333 0.00000000 1.0000000 0.37500000
#> [722,]        NA         NA 1.0000000         NA
#> [723,]        NA         NA        NA 1.00000000
#> [724,]        NA         NA        NA 1.00000000
#> [725,] 1.0000000         NA        NA         NA
#> [726,] 0.0000000         NA        NA         NA
#> [727,] 1.0000000 1.00000000 1.0000000 0.00000000
#> [728,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [729,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [730,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [731,] 1.0000000 0.80000000 1.0000000 1.00000000
#> [732,] 1.0000000 0.75000000 1.0000000 1.00000000
#> [733,] 1.0000000 0.71428571 0.8333333 1.00000000
#> [734,] 1.0000000 1.00000000 0.8000000 1.00000000
#> [735,] 1.0000000 0.75000000 1.0000000 1.00000000
#> [736,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [737,] 1.0000000 0.77777778 0.9230769 1.00000000
#> [738,] 1.0000000 1.00000000 1.0000000 1.00000000
#> [739,] 1.0000000 0.71428571 0.8571429 1.00000000
#> [740,] 1.0000000 0.42857143 1.0000000 1.00000000
#> [741,] 0.6666667 0.16666667 0.8750000 1.00000000
#> [742,] 0.3333333 0.12500000 1.0000000 0.60000000
#> [743,] 1.0000000 0.60000000 1.0000000 1.00000000

#Get methylation matrix along with loci
get_matrix(m = methrix_data, type = 'M', add_loci = TRUE)

#>        chr    start strand        C1        C2        N1    N2
#>   1: chr21 27866423      * 0.1538462 0.2857143 0.5555556 0.300
#>   2: chr21 27866575      *        NA 0.5000000 0.0000000    NA
#>   3: chr21 27866921      * 0.5555556 0.7000000 0.3333333 0.800
#>   4: chr21 27867197      * 0.1818182 0.2500000 0.5833333 0.250
#>   5: chr21 27867248      * 0.6666667 1.0000000 0.8823529 0.875
#>  ---                                                          
#> 739: chr22 49007313      * 1.0000000 0.7142857 0.8571429 1.000
#> 740: chr22 49007329      * 1.0000000 0.4285714 1.0000000 1.000
#> 741: chr22 49007347      * 0.6666667 0.1666667 0.8750000 1.000
#> 742: chr22 49007375      * 0.3333333 0.1250000 1.0000000 0.600
#> 743: chr22 49007398      * 1.0000000 0.6000000 1.0000000 1.000

#' #Get methylation data as a GRanges object
get_matrix(m = methrix_data, type = 'M', add_loci = TRUE, in_granges=TRUE)

#> GRanges object with 743 ranges and 4 metadata columns:
#>         seqnames            ranges strand |        C1        C2        N1
#>            <Rle>         <IRanges>  <Rle> | <numeric> <numeric> <numeric>
#>     [1]    chr21 27866423-27866424      * |  0.153846  0.285714  0.555556
#>     [2]    chr21 27866575-27866576      * |        NA  0.500000  0.000000
#>     [3]    chr21 27866921-27866922      * |  0.555556  0.700000  0.333333
#>     [4]    chr21 27867197-27867198      * |  0.181818  0.250000  0.583333
#>     [5]    chr21 27867248-27867249      * |  0.666667  1.000000  0.882353
#>     ...      ...               ...    ... .       ...       ...       ...
#>   [739]    chr22 49007313-49007314      * |  1.000000  0.714286  0.857143
#>   [740]    chr22 49007329-49007330      * |  1.000000  0.428571  1.000000
#>   [741]    chr22 49007347-49007348      * |  0.666667  0.166667  0.875000
#>   [742]    chr22 49007375-49007376      * |  0.333333  0.125000  1.000000
#>   [743]    chr22 49007398-49007399      * |  1.000000  0.600000  1.000000
#>                N2
#>         <numeric>
#>     [1]     0.300
#>     [2]        NA
#>     [3]     0.800
#>     [4]     0.250
#>     [5]     0.875
#>     ...       ...
#>   [739]       1.0
#>   [740]       1.0
#>   [741]       1.0
#>   [742]       0.6
#>   [743]       1.0
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths