Creates a detailed interative html summary report from Methrix object

Creates a detailed interative html summary report from Methrix object. If the directory contains required files (from previous run), it directly proceeds to generate html report.

methrix_report(
  meth,
  output_dir = NULL,
  recal_stats = FALSE,
  plot_beta_dist = TRUE,
  beta_nCpG = 10000,
  prefix = NULL,
  n_thr = 4
)

Arguments

meth
methrix object

output_dir
Output directory name where the files should be saved. If NULL creats a tempdir

recal_stats
Whether summary statistics should be recalculated? If you are using subsetted methrix object set this to TRUE.

plot_beta_dist
Default TRUE. Can be time consuming.

beta_nCpG
Number of CpGs rto use for estimating beta value distribution. Default 10000

prefix
If provided, the name of the report and the intermediate files will start with the prefix.

n_thr
Default 4. Only used if plot_beta_dist is TRUE

Value

an interactive html report

Examples

if (FALSE) {
data('methrix_data')
methrix::methrix_report(meth = methrix_data)
}