Filter matrices by region

Filter matrices by region

region_filter(m, regions, type = "within")

Arguments

m
methrix object

regions
genomic regions to filter-out. Could be a data.table with 3 columns (chr, start, end) or a GenomicRanges object

type
defines the type of the overlap of the CpG sites with the target regions. Default value is `within`. For detailed description, see the foverlaps function of the data.table package.

Value

An object of class methrix

Details

Takes methrix object and filters CpGs based on supplied regions in data.table or GRanges format

Examples

data('methrix_data')
region_filter(m = methrix_data,
regions = data.table(chr = 'chr21', start = 27867971, end =  27868103))

#> -Removed 4 CpGs

#> -Finished in:  0.015s elapsed (0.014s cpu)

#> An object of class methrix
#>    n_CpGs: 739
#> n_samples: 4
#>     is_h5: FALSE
#> Reference: hg19